+++ iOSClient/Data/NCManageDatabase.swift
@@ -1101,162 +1101,7 @@ class NCManageDatabase: NSObject {
     // MARK: -
     // MARK: Table Share
 
-    @objc func addShare(account: String, home: String, shares: [NKShare]) {
-
-        let realm = try! Realm()
-
-        do {
-            try realm.write {
-
-                for share in shares {
-
-                    let serverUrlPath = home + share.path
-                    guard let serverUrl = NCUtilityFileSystem.shared.deleteLastPath(serverUrlPath: serverUrlPath, home: home) else {
-                        continue
-                    }
-
-                    let object = tableShare()
-
-                    object.account = account
-                    if let fileName = share.path.components(separatedBy: "/").last {
-                        object.fileName = fileName
-                    }
-                    object.serverUrl = serverUrl
-
-                    object.canEdit = share.canEdit
-                    object.canDelete = share.canDelete
-                    object.date = share.date
-                    object.displaynameFileOwner = share.displaynameFileOwner
-                    object.displaynameOwner = share.displaynameOwner
-                    object.expirationDate = share.expirationDate
-                    object.fileParent = share.fileParent
-                    object.fileSource = share.fileSource
-                    object.fileTarget = share.fileTarget
-                    object.hideDownload = share.hideDownload
-                    object.idShare = share.idShare
-                    object.itemSource = share.itemSource
-                    object.itemType = share.itemType
-                    object.label = share.label
-                    object.mailSend = share.mailSend
-                    object.mimeType = share.mimeType
-                    object.note = share.note
-                    object.parent = share.parent
-                    object.password = share.password
-                    object.path = share.path
-                    object.permissions = share.permissions
-                    object.primaryKey = account + " " + String(share.idShare)
-                    object.sendPasswordByTalk = share.sendPasswordByTalk
-                    object.shareType = share.shareType
-                    object.shareWith = share.shareWith
-                    object.shareWithDisplayname = share.shareWithDisplayname
-                    object.storage = share.storage
-                    object.storageId = share.storageId
-                    object.token = share.token
-                    object.uidOwner = share.uidOwner
-                    object.uidFileOwner = share.uidFileOwner
-                    object.url = share.url
-                    object.userClearAt = share.userClearAt
-                    object.userIcon = share.userIcon
-                    object.userMessage = share.userMessage
-                    object.userStatus = share.userStatus
-
-                    realm.add(object, update: .all)
-                }
-            }
-        } catch let error {
-            NKCommon.shared.writeLog("Could not write to database: \(error)")
-        }
-    }
-
-    @objc func getTableShares(account: String) -> [tableShare] {
-
-        let realm = try! Realm()
-
-        let sortProperties = [SortDescriptor(keyPath: "shareType", ascending: false), SortDescriptor(keyPath: "idShare", ascending: false)]
-        let results = realm.objects(tableShare.self).filter("account == %@", account).sorted(by: sortProperties)
-
-        return Array(results.map { tableShare.init(value: $0) })
-    }
-
-    func getTableShares(metadata: tableMetadata) -> (firstShareLink: tableShare?, share: [tableShare]?) {
-
-        let realm = try! Realm()
-
-        let sortProperties = [SortDescriptor(keyPath: "shareType", ascending: false), SortDescriptor(keyPath: "idShare", ascending: false)]
-        let firstShareLink = realm.objects(tableShare.self).filter("account == %@ AND serverUrl == %@ AND fileName == %@ AND shareType == 3", metadata.account, metadata.serverUrl, metadata.fileName).first
-
-        if let firstShareLink = firstShareLink {
-            let results = realm.objects(tableShare.self).filter("account == %@ AND serverUrl == %@ AND fileName == %@ AND idShare != %d", metadata.account, metadata.serverUrl, metadata.fileName, firstShareLink.idShare).sorted(by: sortProperties)
-            return(firstShareLink: tableShare.init(value: firstShareLink), share: Array(results.map { tableShare.init(value: $0) }))
-        } else {
-            let results = realm.objects(tableShare.self).filter("account == %@ AND serverUrl == %@ AND fileName == %@", metadata.account, metadata.serverUrl, metadata.fileName).sorted(by: sortProperties)
-            return(firstShareLink: firstShareLink, share: Array(results.map { tableShare.init(value: $0) }))
-        }
-    }
-
-    func getTableShare(account: String, idShare: Int) -> tableShare? {
-
-        let realm = try! Realm()
-
-        guard let result = realm.objects(tableShare.self).filter("account = %@ AND idShare = %d", account, idShare).first else {
-            return nil
-        }
-
-        return tableShare.init(value: result)
-    }
-
-    @objc func getTableShares(account: String, serverUrl: String) -> [tableShare] {
-
-        let realm = try! Realm()
-
-        let sortProperties = [SortDescriptor(keyPath: "shareType", ascending: false), SortDescriptor(keyPath: "idShare", ascending: false)]
-        let results = realm.objects(tableShare.self).filter("account == %@ AND serverUrl == %@", account, serverUrl).sorted(by: sortProperties)
-
-        return Array(results.map { tableShare.init(value: $0) })
-    }
-
-    @objc func getTableShares(account: String, serverUrl: String, fileName: String) -> [tableShare] {
-
-        let realm = try! Realm()
-
-        let sortProperties = [SortDescriptor(keyPath: "shareType", ascending: false), SortDescriptor(keyPath: "idShare", ascending: false)]
-        let results = realm.objects(tableShare.self).filter("account == %@ AND serverUrl == %@ AND fileName == %@", account, serverUrl, fileName).sorted(by: sortProperties)
-
-        return Array(results.map { tableShare.init(value: $0) })
-    }
-
-    @objc func deleteTableShare(account: String, idShare: Int) {
-
-        let realm = try! Realm()
-
-        realm.beginWrite()
-
-        let result = realm.objects(tableShare.self).filter("account == %@ AND idShare == %d", account, idShare)
-        realm.delete(result)
-
-        do {
-            try realm.commitWrite()
-        } catch let error {
-            NKCommon.shared.writeLog("Could not write to database: \(error)")
-        }
-    }
-
-    @objc func deleteTableShare(account: String) {
-
-        let realm = try! Realm()
-
-        realm.beginWrite()
-
-        let result = realm.objects(tableShare.self).filter("account == %@", account)
-        realm.delete(result)
-
-        do {
-            try realm.commitWrite()
-        } catch let error {
-            NKCommon.shared.writeLog("Could not write to database: \(error)")
-        }
-    }
-
+    
     // MARK: -
     // MARK: Table Tag
